+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](https://semver.org).
 This changelog adheres to [Keep a CHANGELOG](https://keepachangelog.com).
 
 ## [Unreleased]
+
+## [0.47.0] - 2024-04-03
 ### Added
 - (PA-5323) Add RedHat 9 (Power9) platform definition to vanagon
 
@@ -1192,7 +1194,8 @@ on Debian < 8 and needs more work and testing.
 
 ## Versions <= 0.3.9 do not have a change log entry
 
-[Unreleased]: https://github.com/puppetlabs/vanagon/compare/0.46.0...HEAD
+[Unreleased]: https://github.com/puppetlabs/vanagon/compare/0.47.0...HEAD
+[0.47.0]: https://github.com/puppetlabs/vanagon/compare/0.46.0...0.47.0
 [0.46.0]: https://github.com/puppetlabs/vanagon/compare/0.45.0...0.46.0
 [0.45.0]: https://github.com/puppetlabs/vanagon/compare/0.44.0...0.45.0
 [0.44.0]: https://github.com/puppetlabs/vanagon/compare/0.43.0...0.44.0